+++ src/Providers/NominatimServiceProvider.php
@@ -5,23 +5,16 @@ declare(strict_types=1);
 namespace Wimski\LaravelNominatim\Providers;
 
 use Illuminate\Config\Repository as Config;
-use Illuminate\Contracts\Container\BindingResolutionException;
 use Illuminate\Contracts\Container\Container;
 use Illuminate\Support\ServiceProvider;
 use Psr\Http\Client\ClientInterface as HttpClientInterface;
 use Psr\Http\Message\RequestFactoryInterface;
 use Psr\Http\Message\UriFactoryInterface;
-use RuntimeException;
-use Wimski\LaravelNominatim\Config\PackageConfig;
-use Wimski\LaravelNominatim\Enums\ServiceEnum;
+use Wimski\LaravelNominatim\Factories\GeocoderServiceFactory;
 use Wimski\Nominatim\Client;
-use Wimski\Nominatim\Config\LocationIqConfig;
-use Wimski\Nominatim\Config\NominatimConfig;
 use Wimski\Nominatim\Contracts\ClientInterface;
 use Wimski\Nominatim\Contracts\GeocoderServiceInterface;
 use Wimski\Nominatim\Contracts\Transformers\GeocodingResponseTransformerInterface;
-use Wimski\Nominatim\GeocoderServices\LocationIqGeocoderService;
-use Wimski\Nominatim\GeocoderServices\NominatimGeocoderService;
 use Wimski\Nominatim\Transformers\GeocodingResponseTransformer;
 
 class NominatimServiceProvider extends ServiceProvider
@@ -41,16 +34,18 @@ class NominatimServiceProvider extends ServiceProvider
         });
 
         $this->app->singleton(GeocoderServiceInterface::class, function (Container $app): GeocoderServiceInterface {
+            /** @var ClientInterface $client */
+            $client = $this->app->make(ClientInterface::class);
+
+            /** @var GeocodingResponseTransformerInterface $transformer */
+            $transformer = $this->app->make(GeocodingResponseTransformerInterface::class);
+
+            $factory = new GeocoderServiceFactory($client, $transformer);
+
             /** @var Config $config */
             $config = $app->make(Config::class);
 
-            $packageConfig = new PackageConfig($config);
-
-            return match ($packageConfig->getService()->getValue()) {
-                ServiceEnum::NOMINATIM   => $this->makeNominatimService($packageConfig),
-                ServiceEnum::LOCATION_IQ => $this->makeLocationIqService($packageConfig),
-                default                  => throw new RuntimeException('Unreachable statement'),
-            };
+            return $factory->make($config);
         });
     }
 
@@ -59,64 +54,6 @@ class NominatimServiceProvider extends ServiceProvider
         $this->publishes([
             $this->getConfigPath() => config_path('nominatim.php'),
         ]);
-    }
-
-    /**
-     * @param PackageConfig $config
-     * @return NominatimGeocoderService
-     * @throws BindingResolutionException
-     */
-    protected function makeNominatimService(PackageConfig $config): NominatimGeocoderService
-    {
-        /** @var NominatimConfig $serviceConfig */
-        $serviceConfig = $config->getServiceConfig();
-
-        return new NominatimGeocoderService(
-            $this->makeClientInterface(),
-            $this->makeGeocodingResponseTransformerInterface(),
-            $serviceConfig,
-        );
-    }
-
-    /**
-     * @param PackageConfig $config
-     * @return LocationIqGeocoderService
-     * @throws BindingResolutionException
-     */
-    protected function makeLocationIqService(PackageConfig $config): LocationIqGeocoderService
-    {
-        /** @var LocationIqConfig $serviceConfig */
-        $serviceConfig = $config->getServiceConfig();
-
-        return new LocationIqGeocoderService(
-            $this->makeClientInterface(),
-            $this->makeGeocodingResponseTransformerInterface(),
-            $serviceConfig,
-        );
-    }
-
-    /**
-     * @return ClientInterface
-     * @throws BindingResolutionException
-     */
-    protected function makeClientInterface(): ClientInterface
-    {
-        /** @var ClientInterface $client */
-        $client = $this->app->make(ClientInterface::class);
-
-        return $client;
-    }
-
-    /**
-     * @return GeocodingResponseTransformerInterface
-     * @throws BindingResolutionException
-     */
-    protected function makeGeocodingResponseTransformerInterface(): GeocodingResponseTransformerInterface
-    {
-        /** @var GeocodingResponseTransformerInterface $transformer */
-        $transformer = $this->app->make(GeocodingResponseTransformerInterface::class);
-
-        return $transformer;
     }
 
     protected function getHttpClient(): ?HttpClientInterface